Implemented extensive support in Streams plugin for uploading files

+++ platform/plugins/Q/handlers/Q/file/post.php
@@ -9,23 +9,25 @@
  * @class Q/file
  * @method post
  * @param {array} [$params] Parameters that can come from the request
- *   @param {string} [$params.icon.data]  Required if $_FILES is empty. Base64-encoded image data URI - see RFC 2397
- *   @param {string} [$params.icon.path="uploads"] parent path under web dir (see subpath)
- * @param {string} [$params.basename] required name of the file, after the subpath
- *   @param {string} [$params.icon.subpath=""] subpath that should follow the path, to save the image under
+ *   @param {string} [$params.data]  Required if $_FILES is empty. Base64-encoded image data URI - see RFC 2397
+ *   @param {string} [$params.path="uploads"] parent path under web dir (see subpath)
+ *   @param {string} [$params.subpath=""] subpath that should follow the path, to save the image under
+ *   @param {string} [$params.name] override the name of the file, after the subpath
  */
 function Q_file_post($params = null)
 {
 	$p = $params
 		? $params
-		: Q::take($_REQUEST, array('data', 'path', 'subpath', 'merge', 'crop', 'save'));
+		: Q::take($_REQUEST, array('data', 'path', 'subpath'));
 	if (!empty($_FILES)) {
 		$file = reset($_FILES);
-		$tmp = $file['tmp_name'];
-		if (empty($p['data'])) {
-			$p['data'] = file_get_contents($tmp);
+		if ($tmp = $file['tmp_name']) {
+			if (empty($p['data'])) {
+				$p['data'] = file_get_contents($tmp);
+				$p['name'] = $file['name'];
+			}
+			unlink($tmp);
 		}
-		unlink($tmp);
 	} else {
 		if (empty($p['data'])) {
 			throw new Q_Exception_RequiredField(array('field' => 'data'), 'data');
